use upstream instead of variable for fastcgi_pass
If we just hardcoded fastcgi_pass, this would prevent startup if
the name weren't resolvable. Using a variable is a workaround, but
is a little but more work with every query, so using the http upstream
module instead, which also might make it easier to override the upstream
host name.